requirements.txt
langchain 
streamlit
torch
accelerate
sentence_transformers
faiss_cpu
python-dotenv
huggingface_hub
ctransformers
replicate
streamlit_chat